% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cer_dan_peaks-data.R
\docType{data}
\name{cer_dan_peaks}
\alias{cer_dan_peaks}
\title{Imaging mass spectrometry data in Cardinal MSI format of a mouse brain cerebellum
using 1,5-DAN MALDI matrix for analysis of phospholipids}
\format{An object of class \code{MSImageSet} with 926 rows and 19888 columns.}
\source{
self
}
\usage{
data(cer_dan_peaks)
}
\description{
Imaging mass spectrometry data in Cardinal MSI format of a mouse brain cerebellum
using 1,5-DAN MALDI matrix for analysis of phospholipids
}
\examples{
data(cer_dan_peaks)
}
\keyword{datasets}
