% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzIntensBoxPlot.R
\name{mzIntensPlot}
\alias{mzIntensPlot}
\title{plot intensity distribution of an ion in different ROIs}
\usage{
mzIntensPlot(cardinaldata, mz = 885.55, roi_names = c("myrois"))
}
\arguments{
\item{cardinaldata}{MSImageSet containing pixels and intensity data}

\item{roi_names}{ROIs saved in the MSImageSet pixelData using selectROI}

\item{mzs}{single mz value for plotting}
}
\value{
ggplot2 object and display of plot
}
\description{
displays intensty data for a given m/z in different user-defined ROIs. If ROIs are absent, will display the distribution of the signal in the entire data set
}
