% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputFusionIMSinfo.R
\name{outputFusionIMSinfo}
\alias{outputFusionIMSinfo}
\title{Output IMS fusion metadata xml}
\usage{
outputFusionIMSinfo(cardinaldata, filename = "mydata",
  data_label = "my fusion data", IMS_spatial_res = 10)
}
\arguments{
\item{cardinaldata}{a Cardinal MSImageSet of full profile data}

\item{filename}{output file name without extension}

\item{data_label}{a description of the data (i.e. rat brain, 10 um / pixel)}

\item{IMS_spatial_res}{IMS image spatial resolution in microns / pixel}
}
\value{
writes xml in working directory
}
\description{
Outputs an eXtensible Markup Language file based on the IMS input
}
