% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputFusionMicroInfo.R
\name{outputFusionMicroInfo}
\alias{outputFusionMicroInfo}
\title{Output microscopy fusion metadata xml}
\usage{
outputFusionMicroInfo(micro_image_array, filename = "myfusiondata",
  data_label = "fusion label", micro_spatial_res = 1, data_points = NULL)
}
\arguments{
\item{micro_image_array}{an image array [x,y,channel] from the RGB microscopy data loaded into R through the tiff, jpeg, or png library}

\item{filename}{output file name}

\item{data_label}{a description of the data (i.e. rat brain, 1 um / pixel)}

\item{micro_spatial_res}{microscopy image spatial resolution in microns / pixel}
}
\value{
writes xml in working directory
}
\description{
Outputs an eXtensible Markup Language file based on the microscopy input
}
