% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputIMSfusion.R
\name{outputIMSfusion}
\alias{outputIMSfusion}
\title{Output imaging mass spectrometry fusion files}
\usage{
outputIMSfusion(cardinaldata, filename = "myfusiondata",
  data_label = "this is my fusion data", type = c("continuous",
  "peak-picked"), IMS_spatial_res = 100)
}
\arguments{
\item{cardinaldata}{a Cardinal MSImageSet of full profile data}

\item{filename}{output file name without extension. a '_ims_data' and '_ims_info' will be tagged onto files approrpiately}

\item{data_label}{a description of the data(i.e. rat brain, 1 um / pixel)}

\item{type}{continus or peak-picked, continous is full profile, peak-picked is peak picked data but does not currently work (TODO)}
}
\value{
writes outputs in working directory
}
\description{
Outputs a comma seperated .txt file in UNIX format and an XML of metadata for the fusion prototype tool
}
