% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputIMSrepImg.R
\name{outputIMSrepImg}
\alias{outputIMSrepImg}
\title{Output representative image for registration IMS to microscopy registration}
\usage{
outputIMSrepImg(cardinaldata, filename = "myIMSregImage", ...)
}
\arguments{
\item{cardinaldata}{a cardinal MSImageSet or ResultSet}

\item{filename}{output file name, will be extended by '.tif'}

\item{...}{arguments passed to Cardinal's image function mz selection, PCA component selection, etc.}
}
\value{
writes tif of IMS data in working directory for registration
}
\description{
Output representative image for IMS to microscopy registration
}
