% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputMicroFusion.R
\name{outputMicroFusion}
\alias{outputMicroFusion}
\title{Output microscopy fusion files}
\usage{
outputMicroFusion(micro_image_array, filename = "myfusionmicrodata",
  mask_image_array = NULL, data_label = "microscopy data",
  micro_spatial_res = 1)
}
\arguments{
\item{micro_image_array}{an image array [x,y,channel] from the RGB microscopy data loaded into R through the tiff, jpeg, or png library}

\item{filename}{output file name without extension. a '_micro_data.txt' and '_micro_info.xml' will be tagged onto files approrpiately}

\item{data_label}{a description of the data(i.e. rat brain, 1 um / pixel)}

\item{micro_spatial_res}{microscopy image spatial resolution in microns / pixel}
}
\value{
writes outputs in working directory
}
\description{
Outputs a comma seperated .txt file in UNIX format for the fusion prototype tool
}
