% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaPlot.R
\name{pcaPlot}
\alias{pcaPlot}
\title{plot 2D PCA scores in R from Cardinal}
\usage{
pcaPlot(pca_data, components = c(1, 2), roi_names = c("myrois"))
}
\arguments{
\item{pca_data}{an object of class ResultSet}

\item{components}{integer vector indicating which PCs to plot, will only plot the first two supplied values}
}
\value{
ggplot of the PCA scores
}
\description{
uses ggplot2 to plot PC scores with optional ROI color coding
}
