% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMeanSpecs.R
\name{plotMeanSpecs}
\alias{plotMeanSpecs}
\title{Plot mean spectra for ROIs}
\usage{
plotMeanSpecs(cardinaldata, roi_names = c("myrois"), ...)
}
\arguments{
\item{cardinaldata}{MSImageSet}

\item{roi_names}{vector of roi_names, if none are found in the pixelData, only the overall mean will be plotted}

\item{...}{arguments passed to image()}
}
\value{
R plot of the mean spectra
}
\description{
see title
}
