% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocIMS.R
\name{rocIMS}
\alias{rocIMS}
\title{Calculate ROC curve for IMS signals accross regions of interest}
\usage{
rocIMS(cardinaldata, mzs = 687.5, positiveROI = "myROI1",
  negativeROI = "myROI2", undersampling = T)
}
\arguments{
\item{cardinaldata}{MSImageSet containing pixels and intensity data for ROC calculation}

\item{mzs}{vector of mz values to run ROC calculation, or 'all' to run on all mzs in dataset}

\item{positiveROI}{boolean vector in MSImageSet pixelData generated by selectROI function, defines 'cases' in ROC calculation}

\item{negativeROI}{boolean vector in MSImageSet pixelData generated by selectROI function, defines 'control' in ROC calculation, alternatively use 'all' to do the positiveROI vs. all other pixels in the dataset}

\item{subset}{boolean, if one class has more samples than another, randomly subset the longer to match the number of samples in the shorter}
}
\value{
list of pROC analyses sorted descending by area under the ROC curve
}
\description{
Calculates the ROC curve and generates a list of pROC objects based on user defined mz vector
}
