% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectROI.R
\name{selectROI}
\alias{selectROI}
\title{Select ROI using Cardinal wrapper}
\usage{
selectROI(cardinaldata, mz = 885.55)
}
\arguments{
\item{cardinaldata}{an object of class MSImageSet}

\item{mz}{the mz value used for plot and ROI selection}
}
\value{
MSImageSet with selection boolean
}
\description{
Select ROI using Cardinal and return it as a boolean vector in the pixelData for downstream analysis
}
